+++ tests/testthat/test-ui-layout.R
@@ -50,12 +50,14 @@ test_that("cXlayout11", {
 test_that("cXlayout12", {
     check_ui_test(cXlayout12())
 
+    warning('fewer pie charts displayed than expected, does not match web example')
     warning("Colors for drug are different, two pieces are green, but one should be orange")
     warning("The subcharts are in a different layout order compared to web example")
 })
 
 test_that("cXlayout13", {
     check_ui_test(cXlayout13())
+    warning("Warning: bottom plot labels differ from web example")
 })
 
 test_that("cXlayout14", {